+++ cypress/integration/main_page_spec.js
@@ -169,7 +169,7 @@ describe('Dashboard Home Page', () => {
     });
 
     it('modifies the UPDATES section and Granules Errors list as datepicker changes.', () => {
-      cy.route('GET', '/stats?*timestamp__from=1233360000000*').as('stats');
+      cy.intercept('GET', '/stats?*timestamp__from=1233360000000*').as('stats');
 
       cy.get('#Errors').contains('2');
       cy.get('#Collections').contains('1');
@@ -213,7 +213,8 @@ describe('Dashboard Home Page', () => {
       // Cypress only allows one stub per url. We make multiple POST requests to the same
       // elasticsearch endpoint. The fixture here returns a combined response of all the
       // responses for one url, effectively stubbing our elasticsearch searches.
-      cy.route('POST', 'http://example.com/_search/', 'fixture:elasticsearch.json');
+      let fixtureName = 'elasticsearch.json';
+      cy.intercept('POST', 'http://example.com/_search/', (req) => req.reply({ statusCode: 200, fixture: fixtureName }));
       cy.visit('/');
 
       cy.get('.overview-num__wrapper-home > ul#distributionErrors > :nth-child(5)').contains('0');
@@ -228,7 +229,7 @@ describe('Dashboard Home Page', () => {
       cy.get('.overview-num__wrapper-home > ul#distributionSuccesses > :nth-child(2)').contains('9');
       cy.get('.overview-num__wrapper-home > ul#distributionSuccesses > :nth-child(1)').contains('7');
 
-      cy.route('POST', 'http://example.com/_search/', 'fixture:updated_elasticsearch.json');
+      fixtureName = 'updated_elasticsearch.json';
 
       cy.get('[data-cy=startDateTime]').within(() => {
         cy.get('input[name=month]').click().type(1);